% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFunctions.R
\name{expectedNumberOfClusters}
\alias{expectedNumberOfClusters}
\title{Calculate the expected number of clusters from the number of individuals and a concentration parameter}
\usage{
expectedNumberOfClusters(n, a)
}
\arguments{
\item{n}{the number of individuals}

\item{a}{the concentration parameter}
}
\value{
the expected number of clusters
}
\description{
Calculate the expected number of clusters from the number of individuals and a concentration parameter
}
\examples{
expectedNumberOfClusters(100,0.2)
expectedNumberOfClusters(100,0.15)
}
