% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/project.R
\name{ptransform}
\alias{ptransform}
\alias{ptransform.default}
\alias{ptransform.gris}
\alias{ptransform.tbl_df}
\title{Transform coordinates}
\usage{
ptransform(x, ...)

\method{ptransform}{default}(x, ...)

\method{ptransform}{tbl_df}(x, target)

\method{ptransform}{gris}(x, target)
}
\arguments{
\item{x}{}

\item{...}{}

\item{target}{}
}
\description{
Transform coordinates
}

