% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simd.R
\name{simd}
\alias{simd}
\title{Create simple coordinates}
\usage{
simd(n, p = NULL)
}
\arguments{
\item{n}{integer, number of coordinates}

\item{p}{optional, PROJ.4 string to return transformed coordinates}
}
\value{
\code{\link[dplyr]{tbl_df}} with two columns, \code{x} and \code{y}
}
\description{
Create simple coordinates
}
\examples{
set.seed(27)
x <- simd(10)
set.seed(27)
y <- simd(10, "+proj=laea")
}

