% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFunctions.R
\name{simulateChineseRestaurant}
\alias{simulateChineseRestaurant}
\title{Simulate a discrete distribution as in the chinese restaurant problem}
\usage{
simulateChineseRestaurant(num_elements, chi)
}
\arguments{
\item{num_elements}{the number of elements to be grouped}

\item{chi}{the concentration parameter}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Simulate a discrete distribution as in the chinese restaurant problem
}
\examples{
simulateChineseRestaurant(100, 0.2)
simulateChineseRestaurant(100, 0.8)
}
